Extract certificate to external file and add CustomTrust class
