+++ android/repository/src/main/res/raw/signets_cert.crt
@@ -1,0 +1,31 @@
+-----BEGIN CERTIFICATE-----
+MIIFRzCCBC+gAwIBAgIQEisXalGtHwIXQta+Ta6mAzANBgkqhkiG9w0BAQsFADCB
+kDELMAkGA1UEBhMCR0IxGzAZBgNVBAgTEkdyZWF0ZXIgTWFuY2hlc3RlcjEQMA4G
+A1UEBxMHU2FsZm9yZDEaMBgGA1UEChMRQ09NT0RPIENBIExpbWl0ZWQxNjA0BgNV
+BAMTLUNPTU9ETyBSU0EgRG9tYWluIFZhbGlkYXRpb24gU2VjdXJlIFNlcnZlciBD
+QTAeFw0xNzExMDcwMDAwMDBaFw0yMTAxMTYyMzU5NTlaMFcxITAfBgNVBAsTGERv
+bWFpbiBDb250cm9sIFZhbGlkYXRlZDEcMBoGA1UECxMTQ09NT0RPIFNTTCBXaWxk
+Y2FyZDEUMBIGA1UEAwwLKi5ldHNtdGwuY2EwggEiMA0GCSqGSIb3DQEBAQUAA4IB
+DwAwggEKAoIBAQC4XwmsnApIdHleFxh2WQNBWy9l6FgW1zcslmfpFJm9lBFGI0Lx
+sOjoHVu06wC5CQgWadtEkm6VI2ebDujQrpPEW0V6tcxJ0AfDSAS8Uin6cNUAODC0
+l3O4Wb3mvN5Aej6I6tTqlfeaXVXmaKdFbMf6dbWK+PkIg/FnoPDY14PJcvXN6+iV
+XyISWoXdek4ee5pPqyKj2HTQp7RLxeX/Ve3ZTJqYhIc/GDrFOdPNy402oBkcd1nh
+Gju8P7BF1eVt/jrPKzM0Smk/sPHqEr3QBLA3iN7pPaY0SDlnmwbMxsT0UBk3j7V6
+n5thhjnAOUPi9K22P2YWxPLd7kYpFq2OQEw5AgMBAAGjggHTMIIBzzAfBgNVHSME
+GDAWgBSQr2o6lFoL2JDqElZz30O0Oija5zAdBgNVHQ4EFgQUvgstff+NmDbIlgT/
+HmZmV9Vo3JwwDgYDVR0PAQH/BAQDAgWgMAwGA1UdEwEB/wQCMAAwHQYDVR0lBBYw
+FAYIKwYBBQUHAwEGCCsGAQUFBwMCME8GA1UdIARIMEYwOgYLKwYBBAGyMQECAgcw
+KzApBggrBgEFBQcCARYdaHR0cHM6Ly9zZWN1cmUuY29tb2RvLmNvbS9DUFMwCAYG
+Z4EMAQIBMFQGA1UdHwRNMEswSaBHoEWGQ2h0dHA6Ly9jcmwuY29tb2RvY2EuY29t
+L0NPTU9ET1JTQURvbWFpblZhbGlkYXRpb25TZWN1cmVTZXJ2ZXJDQS5jcmwwgYUG
+CCsGAQUFBwEBBHkwdzBPBggrBgEFBQcwAoZDaHR0cDovL2NydC5jb21vZG9jYS5j
+b20vQ09NT0RPUlNBRG9tYWluVmFsaWRhdGlvblNlY3VyZVNlcnZlckNBLmNydDAk
+BggrBgEFBQcwAYYYaHR0cDovL29jc3AuY29tb2RvY2EuY29tMCEGA1UdEQQaMBiC
+CyouZXRzbXRsLmNhgglldHNtdGwuY2EwDQYJKoZIhvcNAQELBQADggEBAA0+MZBG
+kGJvjohYnvqG/e/rdgqqIQxxMeWNKb6xIFXM+OXXnUnnwOWZSgbnfY/0KABq1E3g
+46HDobU19pqo/FStqjBMJJFv9PijNm4naOtHmnFPId5S3vkVMPyDlOHGvUKnzpXS
+RJHjYpgS4G1a1cWISi6LneT6Cvc2B08zSlwCz1z76NuLbRkM+/ZGA2ym6JoXrSRb
+vRdQaCcBEJvap8irsjDnG30ShRtBMR1ksPgqacd+hdPof/ws6jpNl7JOs1ldasKt
+i6o75gQXbbfCG9lcMXz0w2S0+/Wpf0XIYfV05H6/tQ78LTVkXlpdZzIfjGJhEjxE
+iE+LXE3FUVeJdC0=
+-----END CERTIFICATE-----
